+++ src/main.rs
@@ -4,7 +4,6 @@ extern crate zkutil;
 
 use bellman_ce::pairing::bn256::Bn256;
 use clap::Clap;
-use std::fmt;
 use std::fs;
 use std::fs::File;
 use std::path::Path;
@@ -14,30 +13,13 @@ use zkutil::circom_circuit::{
     proof_to_json_file, prove as prove2, proving_key_json_file, r1cs_from_bin_file, r1cs_from_json_file, verification_key_json_file,
     verify as verify2, witness_from_json_file, CircomCircuit, R1CS,
 };
+use zkutil::proof_system::ProofSystem;
 
 /// A tool to work with SNARK circuits generated by circom
 #[derive(Clap)]
 struct Opts {
     #[clap(subcommand)]
     command: SubCommand,
-}
-
-#[derive(Clap, Debug, PartialEq)]
-enum ProofSystem {
-    Groth16,
-    Plonk,
-}
-
-impl str::FromStr for ProofSystem {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "groth16" | "Groth16" => Ok(Self::Groth16),
-            "plonk" | "Plonk" | "PLONK" | "PlonK" => Ok(Self::Plonk),
-            _ => Err("Invalid proof system"),
-        }
-    }
 }
 
 #[derive(Clap)]
@@ -255,7 +237,7 @@ fn generate_verifier(opts: GenerateVerifierOpts) {
     if opts.proof_system == ProofSystem::Plonk {
         unimplemented!();
     }
-    
+
     let params = load_params_file(&opts.params);
     create_verifier_sol_file(&params, &opts.verifier).unwrap();
     println!("Created {}", opts.verifier);
@@ -265,7 +247,7 @@ fn export_keys(opts: ExportKeysOpts) {
     if opts.proof_system == ProofSystem::Plonk {
         unimplemented!();
     }
-    
+
     println!("Exporting {}...", opts.params);
     let params = load_params_file(&opts.params);
     let circuit_file = resolve_circuit_file(opts.circuit);
